Implement the framework of the controller area network.

+++ DevEnvironment/CANMessageGenerator/src/main/java/canmessagegenerator/interfaces/ICANSignalRule.java
@@ -59,12 +59,12 @@ public interface ICANSignalRule extends ICANRule {
     /**
      * Gets the minimum.
      */
-    double getMinimum();
+    int getMinimum();
 
     /**
      * Gets the maximum.
      */
-    double getMaximum();
+    int getMaximum();
 
     /**
      * Gets the unit.
